drivers: add idesc_serial_get_ops function for idesc_serial
