+++ src/drivers/tty/serial/Mybuild
@@ -8,6 +8,7 @@ module serial_dvfs {
 	source "ttys_term.c"
 	source "ttys_dvfs.c"
 	source "idesc_serial_dvfs.c"
+	source "idesc_serial.c"
 
 	@IncludeExport(path="drivers", target_name="ttys.h")
 	source "ttys_dvfs.h"
@@ -22,6 +23,7 @@ module serial_oldfs {
 	source "ttys_processing.c"
 	source "ttys_term.c"
 	source "ttys_oldfs.c"
+	source "idesc_serial_oldfs.c"
 	source "idesc_serial.c"
 
 	@IncludeExport(path="drivers", target_name="ttys.h")
